Added md_helper
