+++ MD-Verlet-Integrator/md_prog.py
@@ -1,6 +1,23 @@
+# A simple Psi 4 script to compute Molecular Dynamics Trajectories
+# Numpy and OS python modules are required
+#
+# MD and Velocity Verlet Algorithms were taken from: :
+# Computational Soft Matter:  From Synthetic Polymers to Proteins, Lecture Notes,
+# Norbert Attig, Kurt Binder, Helmut Grubmuller, Kurt Kremer (Eds.),
+# John von Neumann Institute for Computing, Julich,
+# NIC Series, Vol. 23, ISBN 3-00-012641-4, pp. 1-28, 2004
+#
+# Created by: Leonardo dos Anjos Cunha
+# Date: 05/16/17
+# License: GPL v3.0
+
+
 import numpy as np
 import psi4
 import os
+import sys
+sys.path.append('../')
+import md_helper
 
 psi4.core.set_output_file('md_output.dat',False)
 
@@ -18,16 +35,14 @@ timestep =  5                       # Time step for each iteration in time atomi
 max_md_step = 10000                 # Number of MD iterations
 veloc0 = np.zeros((2,3))            # Numpy array (natoms,3) with inital velocities
 trajec = True                       # Boolean: Save all trajectories in a single xyz file (md_trajectories)  for visualization
-grad_method = 'hf/3-21g'            # QC method (with basis set) for energy and gradient
+grad_method = 'hf/3-21g'            # Method (QC with basis set) for energy and gradient
+int_alg = 'veloc_verlet'            # Algorithm to use as integrator
 opt    = False                      # Optimize geometry using F=ma
 
-# Molecular Dynamics Function
-def molec_dyn(timestep,max_md_step,veloc0,grad_method,trajec = True,opt=False):
+# Molecular Dynamics Main Program
     
     #Get initial positions,velocities, accelerations and forces
-    #timestep *= fs_timeau
-    energy,wfn = psi4.energy(grad_method,return_wfn=True)
-    forces = -np.asarray(psi4.gradient(grad_method,ref_wfn=wfn))
+    energy,forces = md_helper.get_forces(grad_method)
     geom = np.asarray(molec.geometry())
     natoms = molec.natom()
     atom_mass = np.asarray([molec.mass(atom) for atom in range(natoms)])*amu2au
@@ -52,7 +67,7 @@ def molec_dyn(timestep,max_md_step,veloc0,grad_method,trajec = True,opt=False):
             molec.save_xyz_file('md_step_'+str(i)+'.xyz',False)
 
         # Updating positions velocities and accelerations using Velocity Verlet Integrator
-        pos_new,vel_new,accel_new,energy_new = veloc_verlet(timestep,pos,veloc,accel,atom_mass,natoms,grad_method)
+        pos_new,vel_new,accel_new,energy_new = md_helper.integrator(int_alg,timestep,pos,veloc,molec,grad_method)
         pos = pos_new
         if (not opt):
             veloc = vel_new
@@ -60,25 +75,5 @@ def molec_dyn(timestep,max_md_step,veloc0,grad_method,trajec = True,opt=False):
         energy = energy_new
     md_energy.close()
     if trajec:
-        md_trajectories(max_md_step)
+        md_helper.md_trajectories(max_md_step)
     print "Done with Molecular Dynamics Program!"
-
-# Velocity Verlet Integrator
-def veloc_verlet(timestep,pos,vel,accel,atom_mass,natoms,grad_method):
-    vel_new =  vel+0.5*timestep*accel
-    pos_new =  pos+timestep*vel_new
-    molec.set_geometry(psi4.core.Matrix.from_array(pos_new))
-    E,wfn = psi4.energy(grad_method,return_wfn=True)
-    force_new = -np.asarray(psi4.gradient(grad_method,ref_wfn=wfn))
-    accel_new = force_new/(atom_mass.reshape((natoms,1)))
-    vel_new += 0*5*timestep*accel_new
-    return pos_new,vel_new,accel_new,E
-
-def md_trajectories(max_md_step):
-    with open('md_trajectories.xyz','w') as outfile:
-      for i in range(1,max_md_step+1):
-        with open('md_step_'+str(i)+'.xyz')as infile:
-            outfile.write(infile.read())
-    os.system('rm md_step*')
-
-molec_dyn(timestep,max_md_step,veloc0,grad_method,trajec,opt)
